Give the plugin config operations their CRUD names
Each of logging, welcomes and roles had the real behaviour in
Ops::<Plugin>::Configure and a thin, uncalled, earlier draft sitting beside
it under the correctly-CRUD-named Ops::<Plugin>::Settings::Update. Picking
the op by name got you the dead one, and for logging the two even
validated differently.

The CRUD name wins: it now holds the real behaviour and Configure is gone.
Reminders already looked like this — it is the one plugin with no enable
toggle, so it never grew a Configure — and all four now read the same.

The discarded drafts take their specs with them. The logging draft's
channel-visibility warning is not lost functionality: nothing called it,
and that warning is rendered by Components::Logging::ConfigForm and its
channel-warning Stimulus controller.
