--- app/operations/welcomes/configure.rb
@@ -1,41 +1,0 @@
-# frozen_string_literal: true
-
-module Ops
-  module Welcomes
-    class Configure < ApplicationOperation
-      include Ops::PluginConfiguration
-
-      receives :server_configuration,
-        :channel_id,
-        :join_message,
-        :leave_message,
-        :ping_on_join,
-        :suppress_removal_messages,
-        :enabled
-
-      def call
-        settings = server_configuration.welcome_settings
-        settings.assign_attributes(
-          channel_id:,
-          join_message:,
-          leave_message:,
-          ping_on_join:,
-          suppress_removal_messages:
-        )
-        activation = staged_activation
-
-        return failure(messages(settings, activation), value: activation) unless settings.valid? && activation.valid?
-
-        settings.save!
-        save_activation!(activation)
-        ok(activation)
-      end
-
-      private
-
-      def plugin_key
-        :welcomes
-      end
-    end
-  end
-end
