+++ spec/operations/welcomes/settings/update_spec.rb
@@ -4,48 +4,53 @@ require "rails_helper"
 
 RSpec.describe Ops::Welcomes::Settings::Update do
   subject(:result) do
-    described_class.call(server_configuration: server, channel_id:, join_message:, leave_message:)
+    described_class.call(
+      server_configuration: config,
+      channel_id:,
+      join_message: "hi {user}",
+      leave_message: "bye",
+      ping_on_join: false,
+      suppress_removal_messages: true,
+      enabled:
+    )
   end
 
-  let(:server) { create(:server_configuration, discord_id: 1) }
-  let!(:setting) { server.create_welcome_settings! }
-  let(:channel_id) { 99 }
-  let(:join_message) { "hi" }
-  let(:leave_message) { "bye" }
+  let(:config) { create(:server_configuration) }
+  let!(:plugin) { create(:plugin, key: "welcomes", name: "Welcomes") }
+  let!(:settings) { config.create_welcome_settings! }
+  let(:channel_id) { 111 }
+  let(:enabled) { "1" }
 
-  it "sets the channel and messages" do
-    expect(result.success?).to be(true)
-    setting.reload
-    expect(setting.channel_id).to eq(99)
-    expect(setting.join_message).to eq("hi")
-    expect(setting.leave_message).to eq("bye")
-  end
-
-  context "updating existing values" do
-    before do
-      setting.update!(channel_id: 1, join_message: "old")
-    end
-
-    let(:channel_id) { 2 }
-    let(:join_message) { "new" }
-    let(:leave_message) { nil }
-
-    it "overwrites them" do
-      result
-      setting.reload
-      expect(setting.channel_id).to eq(2)
-      expect(setting.join_message).to eq("new")
-      expect(setting.leave_message).to be_nil
+  context "with a channel, enabling the plugin" do
+    it "saves the settings and enables the plugin" do
+      expect(result).to be_success
+      expect(config.welcome_settings.reload.channel_id).to eq(111)
+      expect(config.welcome_settings.join_message).to eq("hi {user}")
+      expect(config.plugin_activations.find_by(plugin:).enabled).to be(true)
+      expect(config.welcome_settings.ping_on_join).to be(false)
+      expect(config.welcome_settings.suppress_removal_messages).to be(true)
     end
   end
 
-  context "without a channel" do
-    let(:channel_id) { nil }
-    let(:leave_message) { nil }
+  context "when enabling without a channel" do
+    let(:channel_id) { "" }
 
-    it "still saves (a channel is required only to enable)" do
-      expect(result.success?).to be(true)
-      expect(setting.reload.channel_id).to be_nil
+    it "fails with an error on the enabled field and persists nothing" do
+      expect(result).to be_failure
+      expect(result.value.errors[:enabled]).to be_present
+      expect(config.welcome_settings.reload.channel_id).to be_nil
+      expect(config.plugin_activations.reload).to be_empty
+    end
+  end
+
+  context "when saving messages without enabling" do
+    let(:enabled) { "0" }
+    let(:channel_id) { "" }
+
+    it "saves the messages and leaves the plugin disabled" do
+      expect(result).to be_success
+      expect(config.welcome_settings.reload.join_message).to eq("hi {user}")
+      expect(config.plugin_activations.find_by(plugin:)&.enabled).to be_falsey
     end
   end
 end
